+ The MBE fully compiles till the molpro interface   + add the code to define fragment names in interaction energy calculations
 On branch manybody_expansion
 Your branch is up-to-date with 'origin/manybody_expansion'.

 Changes to be committed:
	modified:   test/makefile
	modified:   ../iquads/command_parser.hpp
	modified:   ../manybody_expansion/client.cpp
	modified:   ../manybody_expansion/command_setting.hpp
	modified:   ../manybody_expansion/compute_expansion_term_periodic_traits.hpp
	modified:   ../manybody_expansion/compute_interaction_energy_traits.hpp
	modified:   ../manybody_expansion/makefile
	modified:   ../manybody_expansion/polymer_report_subgroup_traits.hpp
	modified:   ../manybody_expansion/test/main.cpp
	modified:   ../manybody_expansion/test/makefile

 Untracked files:
	../../bin/
	../chemistry/test/main.o
	../chemistry/test/test_element
	test/client.o
	test/input/
	test/main.o
	test/molpro_agent.o
	test/output/
	test/report.o
	test/scratch/
	test/setting.o
	test/test_electron_correlation
	../file_system/test/main.o
	../file_system/test/result/
	../file_system/test/test_file_system
	../interface_to_third_party/test/molpro_agent.o
	../interface_to_third_party/test/test_molpro_agent
	../manybody_expansion/agent.o
	../manybody_expansion/client.o
	../manybody_expansion/config.o
	../manybody_expansion/main.o
	../manybody_expansion/mbe
	../manybody_expansion/request.o
	../manybody_expansion/test/.exrc
	../manybody_expansion/test/main.o
	../manybody_expansion/test/request.o
	../manybody_expansion/test/setting.o
	../manybody_expansion/test/test
	../mbe_checkout.sh
	../structure/test/main.o
	../structure/test/test
	../utility/test/main.o
	../utility/test/test_utility
	../../test/nonadiabatic_coupling/
	../../test/numerical_nonadiabatic_coupling/

+++ src/iquads/command_parser.hpp
@@ -51,7 +51,7 @@ public:
   CommandParser( int argc, char* argv[] ){
    cout << "::CommandParser::" << endl;
    cout << "  command arguments:";
-   for( size_t i = 0; i < argc; i++ ){
+   for( int i = 0; i < argc; i++ ){
     const argument_type argument = argv[i];
     this->command_string_.push_back( argument );
     cout << argument;
